move: chatroom dir
