+++ frontend/components/chat/chatroom/ChatroomJoinButton.tsx
@@ -4,7 +4,7 @@ import { Button } from '@mui/material';
 import AddCircleOutlineRounded from '@mui/icons-material/AddCircleOutlineRounded';
 import { CHATROOM_TYPE } from 'types/chat';
 import { User, Chatroom } from '@prisma/client';
-import { ChatroomJoinDialog } from 'components/chat/ChatroomJoinDialog';
+import { ChatroomJoinDialog } from 'components/chat/chatroom/ChatroomJoinDialog';
 
 type Props = {
   socket: Socket;
