+++ frontend/pages/chat/index.tsx
@@ -4,9 +4,9 @@ import { List, TextField, IconButton } from '@mui/material';
 import Grid from '@mui/material/Unstable_Grid2';
 import SendIcon from '@mui/icons-material/Send';
 import { Header } from 'components/common/Header';
-import { ChatroomListItem } from 'components/chat/ChatroomListItem';
-import { ChatroomCreateButton } from 'components/chat/ChatroomCreateButton';
-import { ChatroomJoinButton } from 'components/chat/ChatroomJoinButton';
+import { ChatroomListItem } from 'components/chat/chatroom/ChatroomListItem';
+import { ChatroomCreateButton } from 'components/chat/chatroom/ChatroomCreateButton';
+import { ChatroomJoinButton } from 'components/chat/chatroom/ChatroomJoinButton';
 import { FriendSidebar } from 'components/chat/friend/FriendSidebar';
 import { Chatroom, Message } from 'types/chat';
 import { useQueryUser } from 'hooks/useQueryUser';
